Persistance layer was fixed.
