--- src/main/java/com/solvd/timetable/persistence/SubjectCountRepository.java
@@ -1,10 +1,0 @@
-package com.solvd.timetable.persistence;
-
-import com.solvd.timetable.domain.SubjectCount;
-
-import java.util.List;
-
-public interface SubjectCountRepository {
-
-    List<SubjectCount> getSubjectCounts();
-}
