--- src/main/java/com/solvd/timetable/persistence/SubjectRepository.java
@@ -1,10 +1,0 @@
-package com.solvd.timetable.persistence;
-
-import com.solvd.timetable.domain.Subject;
-
-import java.util.List;
-
-public interface SubjectRepository {
-
-    List<Subject> getSubjects();
-}
