--- src/main/java/com/solvd/timetable/service/SubjectCountService.java
@@ -1,11 +1,0 @@
-package com.solvd.timetable.service;
-
-import com.solvd.timetable.domain.SubjectCount;
-
-import java.util.List;
-
-public interface SubjectCountService {
-
-    List<SubjectCount> getAll();
-
-}
