--- src/main/java/com/solvd/timetable/service/SubjectService.java
@@ -1,11 +1,0 @@
-package com.solvd.timetable.service;
-
-import com.solvd.timetable.domain.Subject;
-
-import java.util.List;
-
-public interface SubjectService {
-
-    List<Subject> getAll();
-
-}
