+++ src/main/resources/mapper/GradeCurriculumMapper.xml
@@ -3,18 +3,35 @@
 <mapper namespace="com.solvd.timetable.persistence.GradeCurriculumRepository">
 
     <select id="getGradeCurricula" resultMap="GradeCurriculumResultMap">
-        Select s.id as subject_id, s.name as subject_name, s.room_type_needed as subject_room_type,
-        spw.id as subject_per_week_id, spw.class_id as subject_per_week_class_id,
-        spw.subject_id as subject_per_week_subject_id, spw.count as subject_per_week_count,
-        c.id as class_id, c.name as class_name
-        from subjects s inner join subject_per_week spw on
-        spw.subject_id = s.id inner join classes c on
-        spw.class_id = c.id
+        Select
+            s.id as subject_id,
+            s.name as subject_name,
+            s.room_type_needed as subject_room_type,
+            spw.id as subject_per_week_id,
+            spw.class_id as subject_per_week_class_id,
+            spw.subject_id as subject_per_week_subject_id,
+            spw.count as subject_per_week_count,
+            c.id as class_id,
+            c.name as class_name
+        from
+            subjects s
+        inner join
+            subject_per_week spw on spw.subject_id = s.id
+        inner join
+            classes c on spw.class_id = c.id
     </select>
 
     <resultMap id="GradeCurriculumResultMap" type="com.solvd.timetable.domain.GradeCurriculum" autoMapping="false">
-        <id property="id" column="subject_per_week_id"/>
+        <id property="id" column="subject_per_week_class_id"/>
         <association property="grade" resultMap="com.solvd.timetable.persistence.GradeRepository.GradeResultMap"/>
-        <association property="countSubjects" resultMap="com.solvd.timetable.persistence.SubjectCountRepository.SubjectCountResultMap"/>
+        <collection property="countSubjects" resultMap="SubjectCountResultMap"/>
     </resultMap>
+
+    <resultMap id="SubjectCountResultMap" type="com.solvd.timetable.domain.SubjectCount" autoMapping="false">
+        <result property="subject.id" column="subject_id"/>
+        <result property="subject.name" column="subject_name"/>
+        <result property="subject.roomTypeNeeded" column="subject_room_type"/>
+        <result property="count" column="subject_per_week_count"/>
+    </resultMap>
+
 </mapper>
