--- src/main/resources/mapper/SubjectCountMapper.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.solvd.timetable.persistence.SubjectCountRepository">
-
-    <select id="getSubjectCounts" resultMap="SubjectCountResultMap">
-        Select s.id as subject_id, s.name as subject_name, s.room_type_needed as subject_room_type,
-        spw.count as subject_per_week_count from subjects s left join subject_per_week
-        on spw.subject_id = s.id
-    </select>
-
-    <resultMap id="SubjectCountResultMap" type="com.solvd.timetable.domain.SubjectCount" autoMapping="false">
-        <result property="count" column="subject_per_week_count"/>
-        <association property="subject" resultMap="com.solvd.timetable.persistence.SubjectRepository.SubjectResultMap"/>
-    </resultMap>
-</mapper>
